Refactor MilkDrop catalog store modules
